+++ application/controllers/manage.php
@@ -7,6 +7,7 @@ class Manage extends CI_Controller {
 		if(!$this->session->userdata('username') && !$this->session->userdata('userid')){
 			redirect('login');
 		}
+		$this->userid = $this->session->userdata('userid');
 		$this->load->model('executive_model');
 		$this->load->model('projects_model');
 		$this->load->model('units_model');
@@ -36,7 +37,7 @@ class Manage extends CI_Controller {
 		$this->form_validation->set_rules('brokerage', 'Brokerage', 'required|xss_safe|numeric');
 
 
-		$this->form_validation->set_rules('type', 'Unit Type', 'required|xss_safe');
+		$this->form_validation->set_rules('type[]', 'Unit Type', 'required|xss_safe');
 
 		$this->form_validation->set_rules('size[]', 'Unit Size', 'required|numeric');
 
@@ -51,9 +52,9 @@ class Manage extends CI_Controller {
 			$array = $this->input->post('size');
 			$type = $this->input->post('type');
 			foreach ($array as $key => $value) {
-				$this->units_model->enter_data($project_id, $type, $key+1, $value);
+				$this->units_model->enter_data($project_id, $type[$key], $key+1, $value);
 			}
-			redirect('manage');
+			redirect('manage/add_property');
 		}
 		
 	}
@@ -94,7 +95,8 @@ class Manage extends CI_Controller {
 		$project = $this->input->post('project');
 		$suggestion = $this->projects_model->search($project);
 		foreach ($suggestion as $key => $value) {
-			echo "<a class='disable' onClick=\"init($value->id)\" value=$value->id href=\"javascript:void(0)\">$value->project</a><br/>";
+			
+			echo "<a class='disable' onClick=init() data-id=\"$value->id\" data-name=\"$value->project\" value=$value->id href=\"javascript:void(0)\">$value->project</a><br/>";
 		}
 	}	
 	public function search_types(){
@@ -111,6 +113,34 @@ class Manage extends CI_Controller {
 			echo "<a class='disable' onClick=init($value->id) value=$value->id href=\"javascript:void(0)\">$value->project</a><br/>";
 		}*/
 	}
+	public function account(){
+		$this->form_validation->set_rules('old_password', 'Old Password', 'required|xss_safe');
+		$this->form_validation->set_rules('new_password', 'New Password', 'required|xss_safe');
+		$this->form_validation->set_rules('conf_new_password', 'Confirm Password', 'required|xss_safe');
+		if($this->form_validation->run()===FALSE){
+			$this->load->view('header');
+			$this->load->view('manage_account');
+			$this->load->view('footer');
+		}else{
+			$o_password = $this->input->post('old_password');
+			$n_password = $this->input->post('new_password');
+			$c_password = $this->input->post('conf_new_password');
+			if($this->check_old_password_validity_for_user($this->userid, $o_password)){
+				if($n_password===$c_password){
+					$this->executive_model->update_password($this->userid, $n_password);
+					$this->session->set_flashdata('success', "<span style='background:yellow;'>Success</span>");
+					redirect('manage/account');
+				}else{
+					$this->session->set_flashdata('error', "<span style='background:yellow;'>Confirm password does not matches</span>");
+					redirect('manage/account');
+				}
+			}else{
+				$this->session->set_flashdata('error', "<span style='background:yellow;'>Old Password is incorrect</span>");
+				redirect('manage/account');
+			}
+		}
+		//echo $this->userid;
+	}
 	public function booking(){
 		$this->form_validation->set_rules('project', 'Project', 'required|xss_safe');
 		$this->form_validation->set_rules('types', 'Unit Type', 'required|xss_safe');
@@ -119,10 +149,9 @@ class Manage extends CI_Controller {
 			$this->load->view('manage_booking');
 			$this->load->view('footer');
 		}else{
-			$project = $this->input->post('project');
+			$project = $this->input->post('project_id');
 			$types = $this->input->post('types');
 			$data['executives'] = $this->executive_model->get_all();
-
 			$data['options'] = $this->units_model->get_options($project, $types);
 			$data['types'] = $types;
 			$data['project'] = $project;
@@ -133,6 +162,7 @@ class Manage extends CI_Controller {
 		
 	}
 	public function make_booking($project){
+
 		$revenue_rate = $this->projects_model->get_revenue($project);
 		$type = $this->input->post('types');
 		//$this->units_model->save_data($project, $revenue_rate);
@@ -153,11 +183,13 @@ class Manage extends CI_Controller {
 		$cashback = $this->input->post('cashback');
 		$exec_array = $this->input->post('executive');
 		$contri_array = $this->input->post('contribution');
-		$booking_id = $this->units_model->save_data($project, $revenue_rate, $type, $options, $client_name, $client_id, $client_address, $client_city, $client_email, $client_contact, $booking_date, $unit_no, $rate, $floor_rise, $plc, $car_park, $car_park_cost, $cashback);
+		$comments = $this->input->post('comments');
+		$booking_id = $this->units_model->save_data($project, $revenue_rate, $type, $options, $client_name, $client_id, $client_address, $client_city, $client_email, $client_contact, $booking_date, $unit_no, $rate, $floor_rise, $plc, $car_park, $car_park_cost, $cashback, $comments);
 		foreach ($exec_array as $key => $value) {
 			$this->executive_model->save_contributions($booking_id, $exec_array[$key], $contri_array[$key]);
 			
 		}
+		$this->session->set_flashdata('booking_done', "<span class='flash'>Booking Success!</span>");
 		redirect('manage/booking');
 		/*echo $project;
 		echo br();
@@ -263,18 +295,67 @@ class Manage extends CI_Controller {
 			$type = $this->input->post('type');
 			if($array && $type){
 				foreach ($array as $key => $value) {
-					$this->units_model->enter_data($project_id, $type, $key+1, $value);
+					$this->units_model->enter_data($project_id, $type[$key], $key+1, $value);
 				}
 			}
 			
-			redirect('manage');
+			redirect("manage/edit_project/$project_id");
 		}
 		
 	}
 	public function delete_unit_value($project_id, $uid){
-		$this->units_model->delete_unit_value($project_id, $uid);
-		$this->session->set_flashdata('unit_deletion', 'Successfully removed unit option');
-		redirect("manage/edit_project/$project_id");
+		if($this->check_admin()){
+			$this->units_model->delete_unit_value($project_id, $uid);
+			$this->session->set_flashdata('unit_deletion', "<span class='flash'>Successfully removed unit option</span>");
+			redirect("manage/edit_project/$project_id");
+		}else{
+			$this->session->set_flashdata('unit_deletion', "<span style='background:yellow;'>You are not authorized for this!</span>");
+			redirect("manage/edit_project/$project_id");
+		}
+	}
+	private function check_old_password_validity_for_user($userid, $password){
+		return $this->executive_model->check_old_password_validity_for_user($userid, $password);
+	}
+	public function delete_project($project_id){
+		if($this->check_admin()){
+			if($this->projects_model->delete_entry($project_id)){
+				$this->session->set_flashdata('project_deleted', "<span style='background:yellow;'>Project Deleted</span>");
+				redirect('manage/add_property');
+			}
+		}else{
+			$this->session->set_flashdata('project_deleted', "<span style='background:yellow;'>You are not authorized for this!</span>");
+			redirect('manage/add_property');
+		}
+	}
+	private function check_admin(){
+		if($this->session->userdata('is_admin')==1){
+			return TRUE;
+		}else{
+			return FALSE;
+		}
+	}
+	public function reporting(){
+		if($this->check_admin()){
+			$this->form_validation->set_rules('year[]', 'Year', 'required');
+			$this->form_validation->set_rules('month[]', 'Month', 'required');
+			$this->form_validation->set_rules('exec_id[]', 'Executive', 'required');
+			if($this->form_validation->run()===FALSE){
+				$data['executives'] = $this->executive_model->get_all();
+				$this->load->view('header');
+				$this->load->view('manage_reporting', $data);
+				$this->load->view('footer');
+			}else{
+				$data['all_data'] = $this->incentives_model->get_all_data();
+				$this->load->view('manage_show_reporting', $data);
+				//var_dump($this->input->post('year'));
+				//var_dump($this->input->post('month'));
+				//var_dump($this->input->post('exec_id'));
+			}
+			
+		}else{
+			$this->session->set_flashdata('authorized', "<span class='flash'>Not Authorized</span>");
+			redirect('manage');
+		}
 	}
 	/*
 	public function add_invoice($booking_id){
